yaml no longer loaded by salt.utils
